@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Sudoku.Domain;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
